% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{add}
\alias{add}
\title{Sum two numbers.}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of x and y.
}
\description{
Sum two numbers.
}
\examples{
add(1, 2)
add(-5, 7)
}
