\name{Example-package}
\alias{Example-package}
\alias{Example}
\docType{package}
\title{
\packageTitle{Example}
}
\description{
\packageDescription{Example}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Example}
\packageIndices{Example}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Example}

Maintainer: \packageMaintainer{Example}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
