% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{subtract}
\alias{subtract}
\title{Subtract the second number from the first.}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The result of \code{x - y}
}
\description{
Subtract the second number from the first.
}
\examples{
subtract(2, 1)
subtract(70, 80)
}
